Fix bug when opening files with spaces in their paths (#15)
* Fix bug when opening files with spaces in their paths

This commit simply wraps the path provided to git with quotes to ensure
that git does not get confused.

* Restore original window setting after disabling blame

This is implemented by populating a table of original options when
enabling blame. Then the blame window is closed again, the settings are
restored in the original window.

Thank you for this amazing plugin!